Performance tuning of MapResultSetConverter and EntityResultSetConverter by caching the objects required for value conversion.

Reference)  
Measurements on my local PC.

- EntityResultSetConverter  
  select 500,000 record.  (SqlQueryTest#testQueryFluentCollectEntityWithPerformance())
  - before : 3.85s , after : 1.81s

- MapResultSetConverter  
  select 500,000 record.  (SqlQueryTest#testQueryFluentCollectWithPerformance())
  - before : 3.21s , after : 1.54s